requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
slack-sdk>=3.25.0
slack-bolt>=1.18.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
aiohttp>=3.9.0
ollama>=0.1.7
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
python-multipart>=0.0.6
httpx>=0.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
html2text>=2020.1.16
chainlit>=1.0.0
supabase>=2.0.0
postgrest>=0.13.0
psycopg2-binary>=2.9.0